+++ KeyboardKitDemoKeyboard/Keyboards/SymbolicKeyboard.swift
@@ -36,7 +36,7 @@ private extension SymbolicKeyboard {
     ]
     
     static var switchAction: KeyboardAction {
-        .switchToKeyboard(.alphabetic(uppercased: false))
+        .keyboardType(.alphabetic(uppercased: false))
     }
 }
 
@@ -44,7 +44,7 @@ private extension Sequence where Iterator.Element == KeyboardActionRow {
     
     func addingSideActions() -> [Iterator.Element] {
         var actions = map { $0 }
-        actions[2].insert(.switchToKeyboard(.numeric), at: 0)
+        actions[2].insert(.keyboardType(.numeric), at: 0)
         actions[2].insert(.none, at: 1)
         actions[2].append(.none)
         actions[2].append(.backspace)